Move Node to co.topl.node sub-package
