--- node-tetra/src/main/scala/co/topl/NodeAppCommandLineArgs.scala
@@ -1,11 +1,0 @@
-package co.topl
-
-import cats.Show
-
-case class NodeAppCommandLineArgs()
-
-object NodeAppCommandLineArgs {
-
-  implicit val showNodeAppCommandLineArgs: Show[NodeAppCommandLineArgs] =
-    Show.show(args => "Args()")
-}
