+++ node-tetra/src/main/scala/co/topl/node/Args.scala
@@ -1,0 +1,16 @@
+package co.topl.node
+
+import cats.Show
+import mainargs._
+
+@main
+case class Args()
+
+object Args {
+
+  implicit val showArgs: Show[Args] =
+    Show.show(args => "Args()")
+
+  implicit val parserArgs: ParserForClass[Args] =
+    ParserForClass[Args]
+}
